+++ src/components/common/modal/Portal.tsx
@@ -1,7 +1,7 @@
 import { createPortal } from 'react-dom';
 
 function Potal({ children }: { children: React.ReactNode }) {
-  return createPortal(<div>{children}</div>, document.getElementById('modal') as HTMLElement);
+  return createPortal(children, document.getElementById('modal') as HTMLElement);
 }
 
 export default Potal;
