+++ src/components/common/modal/emailConfirmModal/index.tsx
@@ -1,18 +1,14 @@
+import { useState } from 'react';
+
 // import { useRouter } from 'next/router';
-import useOnClickOutside from '@/src/hooks/useOnClickOutside';
-import { useRef, useState } from 'react';
 type EmailModalProps = {
   close: () => void;
 };
-const EmailConfirmModal = (close: EmailModalProps) => {
-  const [inputCode, setInputCode] = useState(''); //리엑트 훅 폼으로 사용 변경 필요
-  // const router = useRouter();
-  const ref = useRef<HTMLDivElement>(null);
 
-  useOnClickOutside(ref, () => {
-    console.log('Closing modal'); // 디버깅 메시지 추가
-    close;
-  });
+const EmailConfirmModal = ({ close }: EmailModalProps) => {
+  const [inputCode, setInputCode] = useState(''); //리엑트 훅 폼으로 사용 변경 필요
+
+  // const router = useRouter();
 
   const handleConfirm = () => {
     // if (inputCode === code) {
@@ -25,29 +21,19 @@ const EmailConfirmModal = (close: EmailModalProps) => {
     console.log('클릭');
   };
 
-  //스타일은 임시로 지피티의 도움^3^
   return (
-    <>
-      <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center">
-        <div className="bg-white p-5 rounded">
-          <h2 className="text-xl mb-4">인증 코드 입력</h2>
-          <input
-            type="text"
-            value={inputCode}
-            onChange={(e) => setInputCode(e.target.value)}
-            className="border p-2 mb-4"
-          />
-          <div className="flex justify-end">
-            <button onClick={handleConfirm} className="bg-blue-500 text-white p-2 rounded mr-2">
-              확인
-            </button>
-            <button onClick={handleConfirm} className="bg-gray-500 text-white p-2 rounded">
-              취소
-            </button>
-          </div>
-        </div>
+    <div className="bg-white p-5 rounded">
+      <h2 className="text-xl mb-4">인증 코드 입력</h2>
+      <input type="text" value={inputCode} onChange={(e) => setInputCode(e.target.value)} className="border p-2 mb-4" />
+      <div className="flex justify-end">
+        <button onClick={handleConfirm} className="bg-blue-500 text-white p-2 rounded mr-2">
+          확인
+        </button>
+        <button onClick={close} className="bg-gray-500 text-white p-2 rounded">
+          취소
+        </button>
       </div>
-    </>
+    </div>
   );
 };
