--- src/hooks/useOnClickOutside.ts
@@ -1,19 +1,0 @@
-import { useEffect, RefObject } from 'react';
-
-function useOnClickOutside(ref: RefObject<HTMLElement>, handler: () => void) {
-  useEffect(() => {
-    const listener = (event: MouseEvent) => {
-      if (!ref.current || ref.current.contains(event.target as Node)) {
-        return;
-      }
-      handler();
-    };
-    console.log('ref click');
-    document.addEventListener('mousedown', listener);
-    return () => {
-      document.removeEventListener('mousedown', listener);
-    };
-  }, [ref, handler]);
-}
-
-export default useOnClickOutside;
